Layers for book created, updated

+++ BookRentalApp.Business/BookService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using BookRentalApp.Business.Dto.Book;
+using BookRentalApp.Business.Dto.Category;
 using BookRentalApp.Business.Interface;
 using BookRentalApp.Data.Entity;
 using BookRentalApp.Data.Interface;
@@ -19,48 +20,95 @@ namespace BookRentalApp.Business
             _mapper = mapper;
         }
 
-        public void Add(CreateBookDto book)
+        public ServiceResult<GetBookByIdDto> Add(CreateBookDto bookDto)
         {
-            //
+            var book = _mapper.Map<Book>(bookDto);
+
+            if (book == null)
+            {
+                return ServiceResult<GetBookByIdDto>.Failed(null, "Failed to map book", 400); // 400 - Bad Request
+            }
+
+            _repository.Add(book);
+            var bookDtoResult = _mapper.Map<GetBookByIdDto>(book);
+            return ServiceResult<GetBookByIdDto>.Success(bookDtoResult, "Book added successfully");
         }
 
-        public void Delete(int id)
+        public ServiceResult<GetBookByIdDto> Delete(int id)
         {
+            var book = _repository.GetById(id);
+
+            if (book == null)
+            {
+                return ServiceResult<GetBookByIdDto>.Failed(null, "Book not found", 404); //404 - Not Found
+            }
+
+            var bookDtoResult = _mapper.Map<GetBookByIdDto>(book);
             _repository.Delete(id);
+            return ServiceResult<GetBookByIdDto>.Success(bookDtoResult, "Book deleted successfully");
+
         }
 
-        public List<GetAllBooksDto> GetAll(int page, int pageSize)
+        public ServiceResult<List<GetAllBooksDto>> GetAll(int page, int pageSize)
         {
-            throw new NotImplementedException();
-        }
+            var books = _repository.GetAll(page, pageSize);
 
-        public GetBookByIdDto GetById(int id, bool withCategory = false)
-        {
-            return _mapper.Map<GetBookByIdDto>(_repository.GetById(id, withCategory));
-        }
-
-        public List<GetBookByIdDto> Search(string title, int? categoryId, double? minPrice)
-        {
-            throw new NotImplementedException();
-        }
-
-        public ServiceResult<GetBookByIdDto> Update(int id, UpdateBookDto book)
-        {
-            var b = _repository.GetById(id);
-
-            if(b is null)
+            if (books == null)
             {
-                return ServiceResult<GetBookByIdDto>.Failed(null, "Book not found", 344);
+                return ServiceResult<List<GetAllBooksDto>>.Failed(null, "Failed to retrieve books", 500); //500 - Internal Server Error
             }
 
-            var bo = _repository.Update(id, _mapper.Map<Book>(book));
+            var bookDtosResult = _mapper.Map<List<GetAllBooksDto>>(books);
+            return ServiceResult<List<GetAllBooksDto>>.Success(bookDtosResult, "Books retrieved successfully");
 
-            if(bo is null)
+        }
+
+        public ServiceResult<GetBookByIdDto> GetById(int id, bool withCategory = false)
+        {
+            var book = _repository.GetById(id);
+
+            if (book == null)
             {
-                return ServiceResult<GetBookByIdDto>.Failed(null, "Book not found", 344);
+                return ServiceResult<GetBookByIdDto>.Failed(null, "Book not found", 404); //404 - Not Found
             }
 
-            return ServiceResult<GetBookByIdDto>.Success(_mapper.Map<GetBookByIdDto>(bo));
+            var bookDtoResult = _mapper.Map<GetBookByIdDto>(book);
+            return ServiceResult<GetBookByIdDto>.Success(bookDtoResult, "Book retrieved successfully");
+        }
+
+        public ServiceResult<List<GetBookByIdDto>> Search(string title, string author, string publisher, string ISBN, int? categoryId, double? minPrice, string categoryName, bool isAvailable)
+        {
+            var books = _repository.Search(title, author, publisher, ISBN, categoryId, minPrice, categoryName, isAvailable);
+
+            if (books == null)
+            {
+                return ServiceResult<List<GetBookByIdDto>>.Failed(null, "Failed to retrieve books", 500); //500 - Internal Server Error
+            }
+
+            var bookDtosResult = _mapper.Map<List<GetBookByIdDto>>(books);
+            return ServiceResult<List<GetBookByIdDto>>.Success(bookDtosResult, "Books retrieved successfully");
+
+        }
+
+        public ServiceResult<GetBookByIdDto> Update(int id, UpdateBookDto bookDto)
+        {
+            var book = _repository.GetById(id);
+
+            if (book == null)
+            {
+                return ServiceResult<GetBookByIdDto>.Failed(null, "Book not found", 404); //404 - Not Found
+            }
+
+            var updatedBook = _repository.Update(id, _mapper.Map<Book>(bookDto));
+
+            if (updatedBook == null)
+            {
+                return ServiceResult<GetBookByIdDto>.Failed(null, "Failed to update the book", 500); //500 - Internal Server Error
+            }
+
+            var bookDtoResult = _mapper.Map<GetBookByIdDto>(updatedBook);
+            return ServiceResult<GetBookByIdDto>.Success(bookDtoResult, "Book updated successfully");
         }
     }
+
 }
